OT148-15: OrganizationRepository
- Changed the Intenger to Long

+++ src/main/java/com/alkemy/ong/repository/OrganizationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrganizationRepository extends CrudRepository<OrganizationModel, Integer> {
+public interface OrganizationRepository extends CrudRepository<OrganizationModel, Long> {
 }
